\name{bayescpp-package}
\alias{bayescpp-package}
\alias{bayescpp}
\docType{package}
\title{
\packageTitle{bayescpp}
}
\description{
\packageDescription{bayescpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bayescpp}
\packageIndices{bayescpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bayescpp}

Maintainer: \packageMaintainer{bayescpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
